Jetzt wird der Import für die Post interface aus der App importiert.
Dekomposition für Hooks hinzugefügt.
