+++ src/App.tsx
@@ -8,9 +8,10 @@ import PostList from "./components/PostList";
 import PostFilter from "./components/PostFilter";
 import MyModal from "./components/UI/Modal/MyModal";
 import MyButton from "./components/UI/Button/MyButton";
+import { usePosts } from "./hooks/usePost";
 import { M } from "vite/dist/node/types.d-aGj9QkWt";
 
-interface Post {
+export interface Post {
   id?: number;
   title: string;
   body: string;
@@ -27,20 +28,7 @@ function App() {
 
   const [modal, setModal] = useState(false);
 
-  const sortedPosts = useMemo(() => {
-    if (filter.sort) {
-      return [...posts].sort((a, b) =>
-        a[filter.sort].localeCompare(b[filter.sort]),
-      );
-    }
-    return posts;
-  }, [filter.sort, posts]);
-
-  const sortedAndSearchedPosts = useMemo(() => {
-    return sortedPosts.filter((post) =>
-      post.title.toLowerCase().includes(filter.query.toLowerCase()),
-    );
-  }, [filter.query, sortedPosts]);
+  const sortedAndSearchedPosts = usePosts(posts, filter.sort, filter.query);
 
   const createPost = (newPost: Post) => {
     setPosts([...posts, newPost]);
